Tolerate empty disk name fields in CAB files.
The optional szDiskPrev and szDiskNext fields in the CFHEADER are
empty in cab files I've observed in the wild (particularly those
output by current dev builds of WiX 4), and as far as I can see,
Windows's native <fdi.h> system for extracting cab files doesn't mind
that at all. But libmspack completely rejects any cab file in which
either of those fields is present but empty.

To fix this, I've added a flag to cabd_read_string which indicates
whether the string is allowed to be empty, and set it to 1 for just
those two fields.