CMake: Fix iconv detection
The current implementations check for iconv.h and set HAVE_ICONV_H but
don't set HAVE_ICONV, which is the macro we actually depend on.
In the autotools build system, HAVE_ICONV is set by the `AM_ICONV`
macro, if you have gettext-devel installed (note, this will cause an
autoreconf warning if you don't have gettext-devel installed).

Anyways... to fix this in CMake, I've added FindICONV.cmake module which
is a slightly modified variant of the official CMake iconv detection
module. This one has some additional features including setting the
ICONV_CONST macro as needed.

This commit fixes the cabextract "encoding" test, which was the final
outstanding issue with CMake support, excluding the fact that there
are no cabextract feature tests for Windows. I think this is acceptable
for now, since it's not like there were any before switching to CMake.